refactor(commons): add bem-block-helper and basic-sizes
 - add basic sizes constant: small, medium, large
 - add BEM_PREFIX constant
 - add bem-block-helper - helper to create all set of BEM classes for "my-block"
 - add re-export index for constants
 - refactor Slider component to use new common helper and constants
