+++ src/components/Slider/PlainSlider/SliderRail.scss
@@ -11,21 +11,22 @@
     -webkit-tap-highlight-color: transparent;
     width: 100%;
 
-    // padding: 13px 0px;
-
     &--small {
       border-radius: 2px;
       height: 2px;
+      padding: 6px 0;
     }
 
     &--medium {
       border-radius: 2px;
       height: 4px;
+      padding: 6px 0;
     }
 
     &--large {
       border-radius: 8px;
       height: 8px;
+      padding: 10px 0;
     }
   }
 }
