+++ src/constants/general.js
@@ -1,0 +1,1 @@
+export const BEM_PREFIX = "monday";
